Implement MAT-5389
