+++ svgTransform.js
@@ -1,0 +1,9 @@
+module.exports = {
+  process() {
+    return "module.exports = {};";
+  },
+  getCacheKey() {
+    // The output is always the same.
+    return "svgTransform";
+  },
+};
